+++ frontend/lib/components/Menu/components/ThreadsButton/ThreadsSection/ThreadsSection.tsx
@@ -1,14 +1,14 @@
 import { ChatEntity } from "@/app/chat/[chatId]/types";
 
-import { ChatHistoryItem } from "./ChatHistoryItem/ChatHistoryItem";
-import styles from "./ChatsSection.module.scss";
+import { ThreadItem } from "./ThreadItem/ThreadItem";
+import styles from "./ThreadsSection.module.scss";
 
 type ChatSectionProps = {
   chats: ChatEntity[];
   title: string;
 };
 
-export const ChatsSection = (props: ChatSectionProps): JSX.Element => {
+export const ThreadsSection = (props: ChatSectionProps): JSX.Element => {
   if (props.chats.length === 0) {
     return <></>;
   }
@@ -18,7 +18,7 @@ export const ChatsSection = (props: ChatSectionProps): JSX.Element => {
       <div>{props.title}</div>
       <div className={styles.chats_wrapper}>
         {props.chats.map((chat) => (
-          <ChatHistoryItem key={chat.chat_id} chatHistoryItem={chat} />
+          <ThreadItem key={chat.chat_id} chatHistoryItem={chat} />
         ))}
       </div>
     </div>